Add ability to change quantity of goods by click

+++ app/src/main/java/com/stacktivity/smartwarehouse/data/entities/Item.kt
@@ -8,13 +8,13 @@ import androidx.room.PrimaryKey
 @Entity
 data class Item(@PrimaryKey(autoGenerate = true) var id: Int?,
     @ColumnInfo(name = "item_name") var itemName: String,
-    @ColumnInfo(name = "item_count") var itemCount: Int,
+    @ColumnInfo(name = "item_count") var itemCount: Float,
     @ColumnInfo(name = "quantity_name ") var qtyName: String
 ) {
 
     @Ignore
     constructor(itemName: String,
-                itemCount: Int,
+                itemCount: Float,
                 qtyName: String
     ): this(null, itemName, itemCount, qtyName)
 }
